backend/requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Google AI and Cloud
google-genai>=1.8.0  
google-cloud-aiplatform==1.40.0  

# Web Automation
playwright>=1.48.0  # Browser automation for Computer Use

# Image Processing
Pillow>=10.2.0  # Python 3.13 compatible
requests>=2.31.0  # For downloading images from URLs

# Utilities
python-dotenv==1.0.0  # Environment variable management

# Optional: Development Dependencies 
# pytest>=7.4.3
# black>=23.12.0
# flake8>=6.1.0
